Count number of bingos as points. Placement not implemented yet.

+++ src/main/java/se/artcomputer/f1/bingo/domain/StandingsDto.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public record StandingsDto(List<FanScore> fanScores) {
 
-    public record FanScore(int position, String fanName, int score) {
+    public record FanScore(int position, String fanName, Long score) {
     }
 }
